+++ snowddl/parser/aggregation_policy.py
@@ -1,34 +1,20 @@
 from snowddl.blueprint import (
-    MaskingPolicyBlueprint,
+    AggregationPolicyBlueprint,
+    AggregationPolicyReference,
     Ident,
-    SchemaObjectIdent,
-    NameWithType,
-    DataType,
-    MaskingPolicyReference,
     ObjectType,
+    SchemaObjectIdent,
     build_schema_object_ident,
 )
 from snowddl.parser.abc_parser import AbstractParser, ParsedFile
 
 
 # fmt: off
-masking_policy_json_schema = {
+aggregation_policy_json_schema = {
     "type": "object",
     "properties": {
-        "arguments": {
-            "type": "object",
-            "additionalProperties": {
-                "type": "string"
-            }
-        },
-        "returns": {
-            "type": "string"
-        },
         "body": {
             "type": "string"
-        },
-        "exempt_other_policies": {
-            "type": "boolean"
         },
         "references": {
             "type": "array",
@@ -49,7 +35,7 @@ masking_policy_json_schema = {
                         "minItems": 1
                     }
                 },
-                "required": ["object_type", "object_name", "columns"],
+                "required": ["object_type", "object_name"],
                 "additionalProperties": False
             },
             "minItems": 1
@@ -58,35 +44,31 @@ masking_policy_json_schema = {
             "type": "string"
         }
     },
-    "required": ["arguments", "returns", "body"],
+    "required": ["body"],
     "additionalProperties": False
 }
 # fmt: on
 
 
-class MaskingPolicyParser(AbstractParser):
+class AggregationPolicyParser(AbstractParser):
     def load_blueprints(self):
-        self.parse_schema_object_files("masking_policy", masking_policy_json_schema, self.process_masking_policy)
+        self.parse_schema_object_files("aggregation_policy", aggregation_policy_json_schema, self.process_aggregation_policy)
 
-    def process_masking_policy(self, f: ParsedFile):
-        arguments = [NameWithType(name=Ident(k), type=DataType(t)) for k, t in f.params.get("arguments", {}).items()]
+    def process_aggregation_policy(self, f: ParsedFile):
         references = []
 
         for a in f.params.get("references", []):
-            ref = MaskingPolicyReference(
+            ref = AggregationPolicyReference(
                 object_type=ObjectType[a["object_type"].upper()],
                 object_name=build_schema_object_ident(self.env_prefix, a["object_name"], f.database, f.schema),
-                columns=[Ident(c) for c in a["columns"]],
+                columns=[Ident(c) for c in a.get("columns", [])],
             )
 
             references.append(ref)
 
-        bp = MaskingPolicyBlueprint(
+        bp = AggregationPolicyBlueprint(
             full_name=SchemaObjectIdent(self.env_prefix, f.database, f.schema, f.name),
             body=f.params["body"],
-            arguments=arguments,
-            returns=DataType(f.params["returns"]),
-            exempt_other_policies=f.params.get("exempt_other_policies", False),
             references=references,
             comment=f.params.get("comment"),
         )